Feature/agency artist event authorization

+++ src/main/java/hello/kpop/agency/controller/AgencyController.java
@@ -9,24 +9,37 @@ import hello.kpop.agency.service.AgencyService;
 import hello.kpop.place.dto.SuccessResponseDto;
 import hello.kpop.place.errorHandler.DefaultRes;
 import hello.kpop.place.errorHandler.StatusCode;
+import hello.kpop.user.Role;
+import hello.kpop.user.User;
+import hello.kpop.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Optional;
 
+@Slf4j
 @RestController
 @RequiredArgsConstructor
 public class AgencyController {
 
     private final AgencyRepository agencyRepository;
     private final AgencyService agencyService;
+    private final UserRepository userRepository;
 
     //소속사 등록
-    @PostMapping("/agency") //추후 파라미터에 로그인 한 관리자 정보 필요
-    public ResponseEntity<AgencyResponseDto> saveAgency(@RequestBody AgencyDto requestDto) throws Exception {
+    @PostMapping("/agency")
+    public ResponseEntity<AgencyResponseDto> saveAgency(@RequestBody AgencyDto requestDto, Authentication authentication) throws Exception {
+
+        //토큰이 없는 경우 실행, 즉 로그아웃인 상태에 실행
+        if(authentication == null) {
+            return new ResponseEntity(DefaultRes.res(StatusCode.UNAUTHORIZED, AgencyResponseMessage.UNAUTHORIZED_AGENCY_REGISTER), HttpStatus.UNAUTHORIZED);
+        }
 
         //소속사 중복 등록 방지
         if(agencyRepository.findByAgencyName(requestDto.getAgencyName()).isPresent()) {
@@ -43,7 +56,18 @@ public class AgencyController {
             return new ResponseEntity(DefaultRes.res(StatusCode.BAD_REQUEST, AgencyResponseMessage.AGENCY_NAME_NOT_ENTERED), HttpStatus.BAD_REQUEST);
         }
 
-        return new ResponseEntity(DefaultRes.res(StatusCode.OK, AgencyResponseMessage.AGENCY_REGISTER_SUCCESS, agencyService.saveAgency(requestDto)),HttpStatus.OK);
+        UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+        User user = userRepository.findByUserEmail(userDetails.getUsername()).orElse(null);
+
+        if (user.getUserType() == Role.ADMIN) {
+            //ADMIN 시 실행
+            requestDto.updateRegId(user.getNickname());
+            log.info("유저닉네임:" + user.getNickname());
+            return new ResponseEntity(DefaultRes.res(StatusCode.OK, AgencyResponseMessage.AGENCY_REGISTER_SUCCESS, agencyService.saveAgency(requestDto)),HttpStatus.OK);
+        } else {
+            //ADMIN이 아닐 시 실행
+            return new ResponseEntity(DefaultRes.res(StatusCode.UNAUTHORIZED, AgencyResponseMessage.UNAUTHORIZED_AGENCY_REGISTER), HttpStatus.UNAUTHORIZED);
+        }
     }
 
     //소속사 전체 조회 (아티스트 포함됨)
@@ -74,8 +98,14 @@ public class AgencyController {
     }
 
     //선택한 소속사 정보 수정
-    @PutMapping("/agency/{agencyId}") //추후 파라미터에 로그인 한 관리자 정보 필요
-    public ResponseEntity<AgencyResponseDto> updateAgency(@PathVariable Long agencyId, @RequestBody AgencyDto requestDto) throws Exception {
+    @PutMapping("/agency/{agencyId}")
+    public ResponseEntity<AgencyResponseDto> updateAgency(@PathVariable Long agencyId, @RequestBody AgencyDto requestDto, Authentication authentication) throws Exception {
+
+        //토큰이 없는 경우 실행, 즉 로그아웃인 상태에 실행
+        if(authentication == null) {
+            return new ResponseEntity(DefaultRes.res(StatusCode.UNAUTHORIZED, AgencyResponseMessage.UNAUTHORIZED_AGENCY_UPDATE), HttpStatus.UNAUTHORIZED);
+        }
+
         Optional<Agency> agencyOptional = agencyRepository.findById(agencyId);
 
         //소속사 ID가 일치하는 게 없을 시 실행
@@ -91,18 +121,42 @@ public class AgencyController {
             return new ResponseEntity(DefaultRes.res(StatusCode.NOT_FOUND, AgencyResponseMessage.DELETED_AGENCY), HttpStatus.NOT_FOUND);
         }
 
-        return new ResponseEntity(DefaultRes.res(StatusCode.OK, AgencyResponseMessage.UPDATE_AGENCY_SUCCESS, agencyService.updateAgency(agencyId, requestDto)),HttpStatus.OK);
+        UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+        User user = userRepository.findByUserEmail(userDetails.getUsername()).orElse(null);
+
+        //관리자만 수정가능
+        if(agency.getRegId().equals(user.getNickname())) {
+            return new ResponseEntity(DefaultRes.res(StatusCode.OK, AgencyResponseMessage.UPDATE_AGENCY_SUCCESS, agencyService.updateAgency(agencyId, requestDto)),HttpStatus.OK);
+        } else {
+            return new ResponseEntity(DefaultRes.res(StatusCode.UNAUTHORIZED, AgencyResponseMessage.UNAUTHORIZED_AGENCY_UPDATE), HttpStatus.UNAUTHORIZED);
+        }
     }
 
     //선택한 소속사 삭제
-    @DeleteMapping("/agency/{agencyId}") //추후 파라미터에 로그인 한 관리자 정보 필요
-    public ResponseEntity<SuccessResponseDto> deleteAgency(@PathVariable Long agencyId) throws Exception {
+    @DeleteMapping("/agency/{agencyId}")
+    public ResponseEntity<SuccessResponseDto> deleteAgency(@PathVariable Long agencyId, Authentication authentication) throws Exception {
+
+        //토큰이 없는 경우 실행, 즉 로그아웃인 상태에 실행
+        if(authentication == null) {
+            return new ResponseEntity(DefaultRes.res(StatusCode.UNAUTHORIZED, AgencyResponseMessage.UNAUTHORIZED_AGENCY_DELETE), HttpStatus.UNAUTHORIZED);
+        }
 
         //소속사 ID가 일치하는 게 없을 시 실행
         if(!(agencyRepository.findById(agencyId).isPresent())) {
             return new ResponseEntity(DefaultRes.res(StatusCode.BAD_REQUEST, AgencyResponseMessage.DELETE_AGENCY_FAIL), HttpStatus.BAD_REQUEST);
         }
 
-        return new ResponseEntity(DefaultRes.res(StatusCode.OK, AgencyResponseMessage.DELETE_AGENCY_SUCCESS, agencyService.deleteAgency(agencyId)),HttpStatus.OK);
+        Optional<Agency> agencyOptional = agencyRepository.findById(agencyId);
+        Agency agency = agencyOptional.get();
+
+        UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+        User user = userRepository.findByUserEmail(userDetails.getUsername()).orElse(null);
+
+        //관리자만 수정가능
+        if(agency.getRegId().equals(user.getNickname())) {
+            return new ResponseEntity(DefaultRes.res(StatusCode.OK, AgencyResponseMessage.DELETE_AGENCY_SUCCESS, agencyService.deleteAgency(agencyId)),HttpStatus.OK);
+        } else {
+            return new ResponseEntity(DefaultRes.res(StatusCode.UNAUTHORIZED, AgencyResponseMessage.UNAUTHORIZED_AGENCY_DELETE), HttpStatus.UNAUTHORIZED);
+        }
     }
 }
